# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Rest API Starter Explorer
  description: A proposed openAPI explorer for Rest Api Starter Template
servers:
  - url: /api/v1
    description: Development server
...
paths:
...
  /auth:
    delete:
      tags:
        - Auth Common
      summary: Delete a user permanently
      description: Deletes a user permanently.
      
      responses:
        '200':
          description: Deleted Successfully.
        '400':
          description: Bad Request.
        '403':
          description: Failed to delete user
        '500':
          description: Internal Server Error
      security:
        - bearerAuth: []
...
  /user:
    get:
      tags:
        - User
      summary: Get all the details of a user
      description: Get a list of all the details of a user by providing access token.
      
      responses:
        '200':
          description: User Data Fetched successfully.
        '403':
          description: Failed, Forbidden to fetch user data.
        '500':
          description: Internal Server Error
      security:
        - bearerAuth: []
    patch:
      tags:
        - User
      summary: Edit User Details
      description: Edit data related to the user.
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  example: Shubham
                lastName:
                  type: string
                  example: Raj
                gender:
                  type: string
                  example: Male
                age:
                  type: integer
                  example: 20
              required:
                - firstName
      responses:
        '200':
          description: User Data Fetched successfully.
        '403':
          description: Failed, Forbidden to fetch user data.
        '500':
          description: Internal Server Error
      security:
        - bearerAuth: []
  /comment:
    post:
      tags:
        - Comment
      summary: Creates a new Comment against a Post
      description: Creates a new Comment with User Id from token and Posts against the Post Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  example: "Awesome!"
                postId:
                  type: string
                  example: 5f7c179c6a9b793b44b1cb72
              required:
                - comment
                - postId
      responses:
        '201':
          description: Comment posted successfully.
        '403':
          description: Unable to create comment.
        '500':
          description: Internal Server Error
      security:
        - bearerAuth: []
...
  /post:
    post:
      tags:
        - Post
      summary: Creates a new Post 
      description: Creates a new Post for Current User with or without images
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                caption:
                  type: string
                  example: "Awesome Caption!"
                images:
                  type: string
                  format: binary
              required:
                - caption
      responses:
        '200':
          description: Post created successfully.
        '400':
          description: Failed Validation.[Empty Caption]
        '403':
          description: Unable to create Post.
        '500':
          description: Internal Server Error
      security:
        - bearerAuth: []
    get:
      tags:
        - Post
      summary: Gets Posts of all followed users
      description: Gets Posts of all followed users
      responses:
        '200':
          description: Fetched post successfully.
        '500':
          description: Internal Server Error
      security:
        - bearerAuth: []
  /post/image:
    post:
      tags:
        - Post
      summary: Uploads an image to server ( Need to confirm on the usage )
      description: Uploads an image to server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '200':
          description: Image uploaded successfully.
        '500':
          description: Internal Server Error
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...